+++ stdlib/LinearAlgebra/src/deprecated.jl
@@ -7,6 +7,7 @@ using Base: @deprecate, depwarn
 @deprecate cond(F::LinearAlgebra.LU, p::Integer) cond(convert(AbstractArray, F), p)
 
 # PR #22188
+export cholfact
 @deprecate cholfact!(A::StridedMatrix, uplo::Symbol, ::Type{Val{false}}) cholfact!(Hermitian(A, uplo), Val(false))
 @deprecate cholfact!(A::StridedMatrix, uplo::Symbol) cholfact!(Hermitian(A, uplo))
 @deprecate cholfact(A::StridedMatrix, uplo::Symbol, ::Type{Val{false}}) cholfact(Hermitian(A, uplo), Val(false))
@@ -19,7 +20,7 @@ using Base: @deprecate, depwarn
 @deprecate isposdef!(A::StridedMatrix, UL::Symbol) isposdef!(Hermitian(A, UL))
 
 # bkfact
-import .LinearAlgebra: bkfact, bkfact!
+export bkfact, bkfact!
 function bkfact(A::StridedMatrix, uplo::Symbol, symmetric::Bool = issymmetric(A), rook::Bool = false)
     depwarn(string("`bkfact` with uplo and symmetric arguments is deprecated, ",
         "use `bkfact($(symmetric ? "Symmetric(" : "Hermitian(")A, :$uplo))` instead."),
@@ -33,6 +34,7 @@ function bkfact!(A::StridedMatrix, uplo::Symbol, symmetric::Bool = issymmetric(A
     return bkfact!(symmetric ? Symmetric(A, uplo) : Hermitian(A, uplo), rook)
 end
 
+export lufact!
 @deprecate sqrtm(A::UpperTriangular{T},::Type{Val{realmatrix}}) where {T,realmatrix} sqrtm(A, Val(realmatrix))
 @deprecate lufact(A::AbstractMatrix, ::Type{Val{false}}) lufact(A, Val(false))
 @deprecate lufact(A::AbstractMatrix, ::Type{Val{true}}) lufact(A, Val(true))
@@ -1260,3 +1262,273 @@ end
 @deprecate scale!(C::AbstractMatrix, a::AbstractVector, B::AbstractMatrix) mul!(C, Diagonal(a), B)
 
 Base.@deprecate_binding trace tr
+
+# deprecate lufact to lu
+export lufact
+@deprecate(lufact(S::LU), lu(S))
+@deprecate(lufact(x::Number), lu(x))
+@deprecate(lufact(A::AbstractMatrix{T}) where T, lu(A))
+@deprecate(lufact(A::AbstractMatrix{T}, pivot::Union{Val{false}, Val{true}}) where T, lu(A, pivot))
+@deprecate(lufact(A::Union{AbstractMatrix{T}, AbstractMatrix{Complex{T}}}, pivot::Union{Val{false}, Val{true}} = Val(true)) where {T<:AbstractFloat}, lu(A, pivot))
+
+# deprecate eigfact to eig
+export eigfact
+@deprecate(eigfact(A::StridedMatrix{T}; permute::Bool=true, scale::Bool=true) where T, eig(A; permute=permute, scale=scale))
+@deprecate(eigfact(x::Number), eig(x))
+@deprecate(eigfact(A::AbstractMatrix{TA}, B::AbstractMatrix{TB}) where {TA,TB}, eig(A, B))
+@deprecate(eigfact(A::Number, B::Number), eig(A, B))
+
+@deprecate(eigfact(A::SymTridiagonal{T}) where T, eig(A))
+@deprecate(eigfact(A::SymTridiagonal{T}, irange::UnitRange) where T, eig(A, irange))
+@deprecate(eigfact(A::SymTridiagonal{T}, vl::Real, vu::Real) where T, eig(A, vl, vu))
+
+@deprecate(eigfact(M::Bidiagonal), eig(M))
+@deprecate(eigfact(A::RealHermSymComplexHerm), eig(A))
+@deprecate(eigfact(A::RealHermSymComplexHerm, irange::UnitRange), eig(A, irange))
+@deprecate(eigfact(A::RealHermSymComplexHerm, vl::Real, vh::Real), eig(A, vl, vh))
+
+@deprecate(eigfact(A::AbstractTriangular), eig(A))
+
+@deprecate(eigfact(D::Diagonal; permute::Bool=true, scale::Bool=true), eig(D; permute=permute, scale=scale))
+
+# deprecate schurfact to schur
+export schurfact
+@deprecate(schurfact(A::StridedMatrix{<:BlasFloat}), schur(A))
+@deprecate(schurfact(A::StridedMatrix{T}) where T, schur(A))
+@deprecate(schurfact(A::StridedMatrix{T},B::StridedMatrix{T}) where {T<:BlasFloat}, schur(A))
+@deprecate(schurfact(A::StridedMatrix{TA}, B::StridedMatrix{TB}) where {TA,TB}, schur(A))
+
+# deprecate lqfact to lq
+export lqfact
+@deprecate lqfact(A::StridedMatrix{<:BlasFloat}) lq(A)
+@deprecate lqfact(x::Number) lq(x)
+
+# deprecate qrfact to qr
+export qrfact
+@deprecate(qrfact(x::Number), qr(x))
+@deprecate(qrfact(v::AbstractVector), qr(v))
+@deprecate(qrfact(A::AbstractMatrix{T}) where T, qr(A))
+@deprecate(qrfact(A::AbstractMatrix{T}, arg) where T, qr(A, arg))
+
+# deprecate bkfact to bk
+# bkfact exported in a deprecation above
+@deprecate(bkfact(A::AbstractMatrix{T}, rook::Bool=false) where {T}, bk(A, rook))
+
+# deprecate cholfact to chol
+# cholfact exported in a deprecation above
+@deprecate(cholfact(A::Union{StridedMatrix,RealHermSymComplexHerm{<:Real,<:StridedMatrix}}, ::Val{false}=Val(false)), chol(A, Val(false)))
+@deprecate(cholfact(A::Union{StridedMatrix,RealHermSymComplexHerm{<:Real,<:StridedMatrix}}, ::Val{true}; tol = 0.0), chol(A, Val(true); tol=tol))
+@deprecate(cholfact(x::Number, uplo::Symbol=:U), chol(x, uplo))
+
+# deprecate ldltfact to ldlt
+export ldltfact
+@deprecate(ldltfact(M::SymTridiagonal{T}) where T, ldlt(M))
+
+# deprecate hessfact to hess
+export hessfact
+@deprecate(hessfact(A::StridedMatrix{<:BlasFloat}), hess(A))
+@deprecate(hessfact(A::StridedMatrix{T}) where T, hess(A))
+
+# deprecate lufact! to lu!
+# lufact! exported in a deprecation above
+@deprecate(lufact!(A::StridedMatrix{T}, pivot::Union{Val{false}, Val{true}} = Val(true)) where T<:BlasFloat, lufact!(A, pivot))
+@deprecate(lufact!(A::HermOrSym, pivot::Union{Val{false}, Val{true}} = Val(true)), lu!(A, pivot))
+@deprecate(lufact!(A::StridedMatrix, pivot::Union{Val{false}, Val{true}} = Val(true)), lu!(A, pivot))
+@deprecate(lufact!(A::Tridiagonal{T,V}, pivot::Union{Val{false}, Val{true}} = Val(true)) where {T,V}, lu!(A, pivot))
+
+# deprecate eigfact! to eig!
+export eigfact!
+@deprecate(eigfact!(A::StridedMatrix{T}; permute::Bool=true, scale::Bool=true) where T<:BlasReal, eig!(A; permute=permute, scale=scale))
+@deprecate(eigfact!(A::StridedMatrix{T}; permute::Bool=true, scale::Bool=true) where T<:BlasComplex, eig!(A; permute=permute, scale=scale))
+@deprecate(eigfact!(A::StridedMatrix{T}, B::StridedMatrix{T}) where T<:BlasReal, eig!(A, B))
+@deprecate(eigfact!(A::StridedMatrix{T}, B::StridedMatrix{T}) where T<:BlasComplex, eig!(A, B))
+@deprecate(eigfact!(A::SymTridiagonal{<:BlasReal}), eig!(A))
+@deprecate(eigfact!(A::SymTridiagonal{<:BlasReal}, irange::UnitRange), eig!(A, irange))
+@deprecate(eigfact!(A::SymTridiagonal{<:BlasReal}, vl::Real, vu::Real), eig!(A, vl, vu))
+@deprecate(eigfact!(A::RealHermSymComplexHerm{<:BlasReal,<:StridedMatrix}), eig!(A))
+@deprecate(eigfact!(A::RealHermSymComplexHerm{<:BlasReal,<:StridedMatrix}, irange::UnitRange), eig!(A, irange))
+@deprecate(eigfact!(A::RealHermSymComplexHerm{T,<:StridedMatrix}, vl::Real, vh::Real) where {T<:BlasReal}, eig!(A, vl, vh))
+@deprecate(eigfact!(A::HermOrSym{T,S}, B::HermOrSym{T,S}) where {T<:BlasReal,S<:StridedMatrix}, eig!(A, B))
+@deprecate(eigfact!(A::Hermitian{T,S}, B::Hermitian{T,S}) where {T<:BlasComplex,S<:StridedMatrix}, eig!(A, B))
+
+# deprecate schurfact! to schur!
+export schurfact!
+@deprecate(schurfact!(A::StridedMatrix{T}, B::StridedMatrix{T}) where {T<:BlasFloat}, schur!(A, B))
+@deprecate(schurfact!(A::StridedMatrix{<:BlasFloat}), schur!(A))
+
+# deprecate lqfact! to lq!
+export lqfact!
+@deprecate(lqfact!(A::StridedMatrix{<:BlasFloat}), lq!(A))
+
+# deprecate qrfact! to qr!
+export qrfact!
+@deprecate(qrfact!(A::StridedMatrix{<:BlasFloat}, ::Val{false}), qr!(A, Val(false)))
+@deprecate(qrfact!(A::StridedMatrix{<:BlasFloat}, ::Val{true}), qr!(A, Val(true)))
+@deprecate(qrfact!(A::StridedMatrix{<:BlasFloat}), qr!(A))
+@deprecate(qrfact!(A::StridedMatrix, ::Val{false}), qr!(A, Val(false)))
+@deprecate(qrfact!(A::StridedMatrix, ::Val{true}), qr!(A, Val(true)))
+@deprecate(qrfact!(A::StridedMatrix), qr!(A))
+
+# deprecate cholfact! to chol!
+export cholfact!
+@deprecate(cholfact!(A::RealHermSymComplexHerm, ::Val{false}=Val(false)), chol!(A, Val(false)))
+@deprecate(cholfact!(A::StridedMatrix, ::Val{false}=Val(false)), chol!(A, Val(false)))
+@deprecate(cholfact!(A::RealHermSymComplexHerm{<:BlasReal,<:StridedMatrix}, ::Val{true}; tol = 0.0), chol!(A, Val(true); tol=tol))
+@deprecate(cholfact!(A::RealHermSymComplexHerm{<:Real}, ::Val{true}; tol = 0.0), chol!(A, Val(true); tol=tol))
+@deprecate(cholfact!(A::StridedMatrix, ::Val{true}; tol = 0.0), chol!(A, Val(true); tol=tol))
+
+# deprecate bkfact! to bk!
+export bkfact!
+@deprecate(bkfact!(A::RealHermSymComplexSym{T,S} where {T<:BlasReal,S<:StridedMatrix}, rook::Bool = false), bk!(A, rook))
+@deprecate(bkfact!(A::Hermitian{T,S} where {T<:BlasComplex,S<:StridedMatrix{T}}, rook::Bool = false), bk!(A, rook))
+@deprecate(bkfact!(A::StridedMatrix{<:BlasFloat}, rook::Bool = false), bk!(A, rook))
+
+# deprecate ldltfact! to ldlt!
+export ldltfact!
+@deprecate(ldltfact!(S::SymTridiagonal{T,V}) where {T<:Real,V}, ldlt!(S))
+
+# deprecate hessfact! to hess!
+export hessfact!
+@deprecate(hessfact!(A::StridedMatrix{<:BlasFloat}), hess!(A))
+
+# deprecate svdfact! to svd!
+export svdfact!
+@deprecate(svdfact!(M::Bidiagonal{<:BlasReal}; full::Bool = false, thin::Union{Bool,Nothing} = nothing), svd!(M; full=full, thin=thin))
+@deprecate(svdfact!(A::StridedMatrix{T}; full::Bool = false, thin::Union{Bool,Nothing} = nothing) where T<:BlasFloat, svd!(A; full=full, thin=thin))
+@deprecate(svdfact!(A::StridedMatrix{T}, B::StridedMatrix{T}) where T<:BlasFloat, svd!(A, B))
+@deprecate(svdfact!(A::AbstractTriangular), svd!(A))
+
+# deprecate svdfact to svd
+export svdfact
+@deprecate(svdfact(D::Diagonal), svd(D))
+@deprecate(svdfact(A::StridedVecOrMat{T}; full::Bool = false, thin::Union{Bool,Nothing} = nothing) where T, svd(A; full=full, thin=thin))
+@deprecate(svdfact(x::Number; full::Bool = false, thin::Union{Bool,Nothing} = nothing), svd(x; full=full, thin=thin))
+@deprecate(svdfact(x::Integer; full::Bool = false, thin::Union{Bool,Nothing} = nothing), svd(x; full=full, thin=thin))
+@deprecate(svdfact(A::StridedMatrix{T}, B::StridedMatrix{T}) where {T<:BlasFloat}, svd(A, B))
+@deprecate(svdfact(A::StridedMatrix{TA}, B::StridedMatrix{TB}) where {TA,TB}, svd(A, B))
+@deprecate(svdfact(x::Number, y::Number), svd(x, y))
+@deprecate(svdfact(M::Bidiagonal; full::Bool = false, thin::Union{Bool,Nothing} = nothing), svd(M; full=full, thin=thin))
+@deprecate(svdfact(A::AbstractTriangular), svd(A))
+
+# deprecate transitional decomposition getindex methods out of the blocks
+@inline function Base.getindex(S::LU, i::Integer)
+    depwarn(string("decomposition functions (e.g. `lu`) now return decomposition ",
+        "objects (e.g. `LU`), and indexing such objects is deprecated. Instead ",
+        "extract components via their accessors (e.g. `X.L`, `X.S`, and `X.p` for ",
+        "`X::LU`), or destructure the decomposition via iteration ",
+        "(e.g. `l, u, p = X` for `X::LU`)."), :getindex)
+    i == 1 ? (return S.L) :
+    i == 2 ? (return S.U) :
+    i == 3 ? (return S.p) :
+        throw(BoundsError(S, i))
+end
+@inline function Base.getindex(S::Union{Eigen,GeneralizedEigen}, i::Integer)
+    depwarn(string("decomposition functions (e.g. `eig`) now return decomposition ",
+        "objects (e.g. `Eigen` and `GeneralizedEigen`), and indexing such objects ",
+        "is deprecated. Instead extract components via their accessors ",
+        "(e.g. `X.values` and `X.vectors` for `X::Union{Eigen,GeneralizedEigen}`), ",
+        "or destructure the decomposition via iteration ",
+        "(e.g. `vals, vecs = X` for `X::Union{Eigen,GeneralizedEigen}`)."), :getindex)
+    i == 1 ? (return S.values) :
+    i == 2 ? (return S.vectors) :
+        throw(BoundsError(S, i))
+end
+@inline function Base.getindex(S::Schur, i::Integer)
+    depwarn(string("decomposition functions (e.g. `schur`) now return decomposition ",
+        "objects (e.g. `Schur`), and indexing such objects ",
+        "is deprecated. Instead extract components via their accessors ",
+        "(e.g. `X.T`, `X.Z`, and `X.values` for `X::Schur`), ",
+        "or destructure the decomposition via iteration ",
+        "(e.g. `t, z, vals = X` for `X::Schur`)."), :getindex)
+    i == 1 ? (return S.T) :
+    i == 2 ? (return S.Z) :
+    i == 3 ? (return S.values) :
+        throw(BoundsError(S, i))
+end
+@inline function Base.getindex(S::GeneralizedSchur, i::Integer)
+    depwarn(string("decomposition functions (e.g. `schur`) now return decomposition ",
+        "objects (e.g. `GeneralizedSchur`), and indexing such objects ",
+        "is deprecated. Instead extract components via their accessors ",
+        "(e.g. `X.S`, `X.T`, `X.Q`, `X.Z`, `X.α`, and `X.β` for `X::GeneralizedSchur`), ",
+        "or destructure the decomposition via iteration ",
+        "(e.g. `s, t, q, z, α, β = X` for `X::GeneralizedSchur`)."), :getindex)
+    i == 1 ? (return S.S) :
+    i == 2 ? (return S.T) :
+    i == 3 ? (return S.Q) :
+    i == 4 ? (return S.Z) :
+    i == 5 ? (return S.α) :
+    i == 6 ? (return S.β) :
+        throw(BoundsError(S, i))
+end
+@inline function Base.getindex(S::LQ, i::Integer)
+    depwarn(string("decomposition functions (e.g. `lq`) now return decomposition ",
+        "objects (e.g. `LQ`), and indexing such objects ",
+        "is deprecated. Instead extract components via their accessors ",
+        "(e.g. `X.L` and `X.Q` for `X::LQ`), ",
+        "or destructure the decomposition via iteration ",
+        "(e.g. `l, q = X` for `X::LQ`)."), :getindex)
+    i == 1 ? (return S.L) :
+    i == 2 ? (return S.Q) :
+        throw(BoundsError(S, i))
+end
+@inline function Base.getindex(S::QR, i::Integer)
+    depwarn(string("decomposition functions (e.g. `qr`) now return decomposition ",
+        "objects (e.g. `QR`), and indexing such objects ",
+        "is deprecated. Instead extract components via their accessors ",
+        "(e.g. `X.Q` and `X.R` for `X::QR`), ",
+        "or destructure the decomposition via iteration ",
+        "(e.g. `q, r = X` for `X::QR`)."), :getindex)
+    i == 1 ? (return S.Q) :
+    i == 2 ? (return S.R) :
+        throw(BoundsError(S, i))
+end
+@inline function Base.getindex(S::QRCompactWY, i::Integer)
+    depwarn(string("decomposition functions (e.g. `qr`) now return decomposition ",
+        "objects (e.g. `QRCompactWY`), and indexing such objects ",
+        "is deprecated. Instead extract components via their accessors ",
+        "(e.g. `X.Q` and `X.R` for `X::QR`), ",
+        "or destructure the decomposition via iteration ",
+        "(e.g. `q, r = X` for `X::QR`)."), :getindex)
+    i == 1 ? (return S.Q) :
+    i == 2 ? (return S.R) :
+        throw(BoundsError(S, i))
+end
+@inline function Base.getindex(S::QRPivoted, i::Integer)
+    depwarn(string("decomposition functions (e.g. `qr`) now return decomposition ",
+        "objects (e.g. `QRPivoted`), and indexing such objects ",
+        "is deprecated. Instead extract components via their accessors ",
+        "(e.g. `X.Q`, `X.R`, and `X.p` for `X::QRPivoted`), ",
+        "or destructure the decomposition via iteration ",
+        "(e.g. `q, r, p = X` for `X::QRPivoted`)."), :getindex)
+    i == 1 ? (return S.Q) :
+    i == 2 ? (return S.R) :
+    i == 3 ? (return S.p) :
+        throw(BoundsError(S, i))
+end
+@inline function Base.getindex(S::SVD, i::Integer)
+    depwarn(string("decomposition functions (e.g. `svd`) now return decomposition ",
+        "objects (e.g. `SVD`), and indexing such objects ",
+        "is deprecated. Instead extract components via their accessors ",
+        "(e.g. `X.U`, `X.S`, and `X.V` for `X::SVD`), ",
+        "or destructure the decomposition via iteration ",
+        "(e.g. `u, s, v = X` for `X::SVD`)."), :getindex)
+    i == 1 ? (return S.U) :
+    i == 2 ? (return S.S) :
+    i == 3 ? (return S.V) :
+        throw(BoundsError(S, i))
+end
+@inline function Base.getindex(S::GeneralizedSVD, i::Integer)
+    depwarn(string("decomposition functions (e.g. `svd`) now return decomposition ",
+        "objects (e.g. `GeneralizedSVD`), and indexing such objects ",
+        "is deprecated. Instead extract components via their accessors ",
+        "(e.g. `X.U`, `X.V`, `X.Q`, `X.D1`, `X.D2`, and `X.R0` for `X::GeneralizedSVD`), ",
+        "or destructure the decomposition via iteration ",
+        "(e.g. `u, v, q, d1, d2, r0 = X` for `X::GeneralizedSVD`)."), :getindex)
+    i == 1 ? (return S.U) :
+    i == 2 ? (return S.V) :
+    i == 3 ? (return S.Q) :
+    i == 4 ? (return S.D1) :
+    i == 5 ? (return S.D2) :
+    i == 6 ? (return S.R0) :
+        throw(BoundsError(S, i))
+end