+++ stdlib/LinearAlgebra/test/svd.jl
@@ -5,7 +5,7 @@ module TestSVD
 using Test, LinearAlgebra, Random
 using LinearAlgebra: BlasComplex, BlasFloat, BlasReal, QRPivoted
 
-@testset "Simple svdvals / svdfact tests" begin
+@testset "Simple svdvals / svd tests" begin
     ≊(x,y) = isapprox(x,y,rtol=1e-15)
 
     m1 = [2 0; 0 0]
@@ -15,8 +15,8 @@ using LinearAlgebra: BlasComplex, BlasFloat, BlasReal, QRPivoted
     @test @inferred(svdvals(m2))  ≊ [2, 1]
     @test @inferred(svdvals(m2c)) ≊ [2, 1]
 
-    sf1 = svdfact(m1)
-    sf2 = svdfact(m2)
+    sf1 = svd(m1)
+    sf2 = svd(m2)
     @test sf1.S ≊ [2, 0]
     @test sf2.S ≊ [2, 1]
     # U & Vt are unitary
@@ -52,7 +52,7 @@ a2img  = randn(n,n)/2
     for (a, a2) in ((aa, aa2), (view(aa, 1:n, 1:n), view(aa2, 1:n, 1:n)))
         ε = εa = eps(abs(float(one(eltya))))
 
-        usv = svdfact(a)
+        usv = svd(a)
         @testset "singular value decomposition" begin
             @test usv.S === svdvals(usv)
             @test usv.U * (Diagonal(usv.S) * usv.Vt) ≈ a
@@ -63,7 +63,7 @@ a2img  = randn(n,n)/2
             @test usv\b ≈ a\b
 
             if eltya <: BlasFloat
-                svdz = svdfact!(Matrix{eltya}(undef,0,0))
+                svdz = svd!(Matrix{eltya}(undef,0,0))
                 @test svdz.U ≈ Matrix{eltya}(I, 0, 0)
                 @test svdz.S ≈ real(zeros(eltya,0))
                 @test svdz.Vt ≈ Matrix{eltya}(I, 0, 0)
@@ -71,7 +71,7 @@ a2img  = randn(n,n)/2
         end
         @testset "Generalized svd" begin
             a_svd = a[1:n1, :]
-            gsvd = svdfact(a,a_svd)
+            gsvd = svd(a,a_svd)
             @test gsvd.U*gsvd.D1*gsvd.R*gsvd.Q' ≈ a
             @test gsvd.V*gsvd.D2*gsvd.R*gsvd.Q' ≈ a_svd
             @test usv.Vt' ≈ usv.V
@@ -79,7 +79,7 @@ a2img  = randn(n,n)/2
             @test_throws ErrorException gsvd.Z
             @test gsvd.vals ≈ svdvals(a,a_svd)
             α = eltya == Int ? -1 : rand(eltya)
-            β = svdfact(α)
+            β = svd(α)
             @test β.S == [abs(α)]
             @test svdvals(α) == abs(α)
             u,v,q,d1,d2,r0 = svd(a,a_svd)
@@ -93,7 +93,7 @@ a2img  = randn(n,n)/2
             #testing the other layout for D1 & D2
             b = rand(eltya,n,2*n)
             c = rand(eltya,n,2*n)
-            gsvd = svdfact(b,c)
+            gsvd = svd(b,c)
             @test gsvd.U*gsvd.D1*gsvd.R*gsvd.Q' ≈ b
             @test gsvd.V*gsvd.D2*gsvd.R*gsvd.Q' ≈ c
         end
@@ -101,18 +101,16 @@ a2img  = randn(n,n)/2
     if eltya <: LinearAlgebra.BlasReal
         @testset "Number input" begin
             x, y = randn(eltya, 2)
-            @test svdfact(x)    == svdfact(fill(x, 1, 1))
+            @test svd(x)    == svd(fill(x, 1, 1))
             @test svdvals(x)    == first(svdvals(fill(x, 1, 1)))
-            @test svd(x)        == first.(svd(fill(x, 1, 1)))
-            @test svdfact(x, y) == svdfact(fill(x, 1, 1), fill(y, 1, 1))
+            @test svd(x, y) == svd(fill(x, 1, 1), fill(y, 1, 1))
             @test svdvals(x, y) ≈  first(svdvals(fill(x, 1, 1), fill(y, 1, 1)))
-            @test svd(x, y)     == first.(svd(fill(x, 1, 1), fill(y, 1, 1)))
         end
     end
     if eltya != Int
         @testset "isequal, ==, and hash" begin
             x, y   = rand(eltya), convert(eltya, NaN)
-            Fx, Fy = svdfact(x), svdfact(y)
+            Fx, Fy = svd(x), svd(y)
             @test   Fx == Fx
             @test !(Fy == Fy)
             @test isequal(Fy, Fy)